requirements.txt
python-telegram-bot==20.7
torch
torchvision
Pillow
urllib3
